# Use IndexAny instead of Split to reduce memory allocation 