benchcmp result:
```
benchcmp is deprecated in favor of benchstat: https://pkg.go.dev/golang.org/x/perf/cmd/benchstat
benchmark                                  old ns/op     new ns/op     delta
BenchmarkRealIPForHeaderXForwardFor-12     128           39.7          -68.98%
```

before:
```
goos: darwin
goarch: amd64
pkg: github.com/labstack/echo/v4
BenchmarkRealIP-12       8774244               130 ns/op              48 B/op          1 allocs/op
PASS
ok      github.com/labstack/echo/v4     1.297s
```

after:
```
goos: darwin
goarch: amd64
pkg: github.com/labstack/echo/v4
BenchmarkRealIP-12      28132903                38.2 ns/op             0 B/op          0 allocs/op
PASS
ok      github.com/labstack/echo/v4     1.133s
```