@lammel 2020-09-15
The title says IndexByte but in the code string.IndexAny() is used.
IndexByte cam with Go 1.2 so that is also fine. Tests should be adjusted to cover the missing cases (as code coverage dropped)

What was used in the benchmarks actually?