@lammel 2020-11-05
Tested for benchmark regressions on the PR branch.

PR branch feature/opt-split:

```
sh# /usr/bin/go test -benchmem -run=^$ github.com/labstack/echo/v4 -bench ^Benchmark -benchtime 5s
goos: linux
goarch: amd64
pkg: github.com/labstack/echo/v4
BenchmarkBindbindData-8                 	  773809	      8006 ns/op	     368 B/op	      40 allocs/op
BenchmarkBindbindDataWithTags-8         	  600234	      9815 ns/op	     408 B/op	      39 allocs/op
BenchmarkAllocJSONP-8                   	16191808	       369 ns/op	     155 B/op	       4 allocs/op
BenchmarkAllocJSON-8                    	24182094	       240 ns/op	      95 B/op	       1 allocs/op
BenchmarkAllocXML-8                     	 4307667	      1398 ns/op	    4751 B/op	      10 allocs/op
BenchmarkRealIPForHeaderXForwardFor-8   	152748471	        38.7 ns/op	       0 B/op	       0 allocs/op
BenchmarkContext_Store-8                	133940372	        44.3 ns/op	       0 B/op	       0 allocs/op
BenchmarkRouterStaticRoutes-8           	  484891	     12681 ns/op	       0 B/op	       0 allocs/op
BenchmarkRouterGitHubAPI-8              	  249950	     23171 ns/op	       0 B/op	       0 allocs/op
BenchmarkRouterParseAPI-8               	 1406133	      4247 ns/op	       0 B/op	       0 allocs/op
BenchmarkRouterGooglePlusAPI-8          	  884396	      6860 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/labstack/echo/v4	81.256s
```

compared to current master:

```
 /usr/bin/go test -benchmem -run=^$ github.com/labstack/echo/v4 -bench ^Benchmark -benchtime 5s
goos: linux
goarch: amd64
pkg: github.com/labstack/echo/v4
BenchmarkBindbindData-8                 	  763737	      8278 ns/op	     368 B/op	      40 allocs/op
BenchmarkBindbindDataWithTags-8         	  599168	      9928 ns/op	     408 B/op	      39 allocs/op
BenchmarkAllocJSONP-8                   	16310584	       377 ns/op	     154 B/op	       4 allocs/op
BenchmarkAllocJSON-8                    	24018079	       244 ns/op	      95 B/op	       1 allocs/op
BenchmarkAllocXML-8                     	 4306189	      1411 ns/op	    4751 B/op	      10 allocs/op
BenchmarkRealIPForHeaderXForwardFor-8   	49710991	       114 ns/op	      48 B/op	       1 allocs/op
BenchmarkContext_Store-8                	132688436	        47.3 ns/op	       0 B/op	       0 allocs/op
BenchmarkRouterStaticRoutes-8           	  539204	     11651 ns/op	       0 B/op	       0 allocs/op
BenchmarkRouterGitHubAPI-8              	  259131	     23709 ns/op	       0 B/op	       0 allocs/op
BenchmarkRouterParseAPI-8               	 1487281	      4015 ns/op	       0 B/op	       0 allocs/op
BenchmarkRouterGooglePlusAPI-8          	  919380	      6719 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/labstack/echo/v4	78.373s
```

No regressions seen. PR looks OK.
Thanks a lot @imxyb . Merged.